feat/cardUIUX unfinished
