+++ frontend/pages/_components/sales-card-skeleton.jsx
@@ -1,0 +1,101 @@
+import { Card, Skeleton } from "@heroui/react";
+
+export const SalesCardSkeleton = () => {
+  const test = "foobar";
+  console.log(test);
+
+  return (
+    <>
+      <Card className="w-[200px] space-y-5 p-4" radius="lg">
+        <div className="space-y-3">
+          <Skeleton className="w-3/5 rounded-lg">
+            <div className="h-3 w-3/5 rounded-lg bg-default-200" />
+          </Skeleton>
+          <Skeleton className="w-4/5 rounded-lg">
+            <div className="h-3 w-4/5 rounded-lg bg-default-200" />
+          </Skeleton>
+        </div>
+        <Skeleton className="rounded-lg">
+          <div className="h-24 rounded-lg bg-default-300" />
+        </Skeleton>
+        <div className="flex justify-end">
+          <Skeleton className="w-1/3 rounded-lg">
+            <div className="bg-default-300 h-3" />
+          </Skeleton>
+        </div>
+      </Card>
+      <Card className="w-[200px] space-y-5 p-4" radius="lg">
+        <div className="space-y-3">
+          <Skeleton className="w-3/5 rounded-lg">
+            <div className="h-3 w-3/5 rounded-lg bg-default-200" />
+          </Skeleton>
+          <Skeleton className="w-4/5 rounded-lg">
+            <div className="h-3 w-4/5 rounded-lg bg-default-200" />
+          </Skeleton>
+        </div>
+        <Skeleton className="rounded-lg">
+          <div className="h-24 rounded-lg bg-default-300" />
+        </Skeleton>
+        <div className="flex justify-end">
+          <Skeleton className="w-1/3 rounded-lg">
+            <div className="bg-default-300 h-3" />
+          </Skeleton>
+        </div>
+      </Card>
+      <Card className="w-[200px] space-y-5 p-4" radius="lg">
+        <div className="space-y-3">
+          <Skeleton className="w-3/5 rounded-lg">
+            <div className="h-3 w-3/5 rounded-lg bg-default-200" />
+          </Skeleton>
+          <Skeleton className="w-4/5 rounded-lg">
+            <div className="h-3 w-4/5 rounded-lg bg-default-200" />
+          </Skeleton>
+        </div>
+        <Skeleton className="rounded-lg">
+          <div className="h-24 rounded-lg bg-default-300" />
+        </Skeleton>
+        <div className="flex justify-end">
+          <Skeleton className="w-1/3 rounded-lg">
+            <div className="bg-default-300 h-3" />
+          </Skeleton>
+        </div>
+      </Card>
+      <Card className="w-[200px] space-y-5 p-4" radius="lg">
+        <div className="space-y-3">
+          <Skeleton className="w-3/5 rounded-lg">
+            <div className="h-3 w-3/5 rounded-lg bg-default-200" />
+          </Skeleton>
+          <Skeleton className="w-4/5 rounded-lg">
+            <div className="h-3 w-4/5 rounded-lg bg-default-200" />
+          </Skeleton>
+        </div>
+        <Skeleton className="rounded-lg">
+          <div className="h-24 rounded-lg bg-default-300" />
+        </Skeleton>
+        <div className="flex justify-end">
+          <Skeleton className="w-1/3 rounded-lg">
+            <div className="bg-default-300 h-3" />
+          </Skeleton>
+        </div>
+      </Card>
+      <Card className="w-[200px] space-y-5 p-4" radius="lg">
+        <div className="space-y-3">
+          <Skeleton className="w-3/5 rounded-lg">
+            <div className="h-3 w-3/5 rounded-lg bg-default-200" />
+          </Skeleton>
+          <Skeleton className="w-4/5 rounded-lg">
+            <div className="h-3 w-4/5 rounded-lg bg-default-200" />
+          </Skeleton>
+        </div>
+        <Skeleton className="rounded-lg">
+          <div className="h-24 rounded-lg bg-default-300" />
+        </Skeleton>
+        <div className="flex justify-end">
+          <Skeleton className="w-1/3 rounded-lg">
+            <div className="bg-default-300 h-3" />
+          </Skeleton>
+        </div>
+      </Card>
+    </>
+  );
+};
